move api file into api directory
